refactor(context-menu): H.2 — extract AddToPlaylistSubmenu component
